Put back the notfifs

+++ views/backoffice/stocks.php
@@ -23,7 +23,7 @@
         // ===============================
         // Suppression notif après clic
         // ===============================
-        /* if (isset($_GET['idNotif'], $_GET['reassort_id'])) {
+        if (isset($_GET['idNotif'], $_GET['reassort_id'])) {
             $del = $pdo->prepare("
                 DELETE FROM _notification 
                 WHERE idNotif = :idNotif 
@@ -49,8 +49,8 @@
         if (isset($_GET['reassort_id'])) {
             $_SESSION['hide_notif'] = true;
         }
-    */?>
-    <?php // require_once './partials/notifications_stock.php' ?>
+    ?>
+    <?php require_once './partials/notifications_stock.php' ?>
 
     <?php
         $currentPage = basename(__FILE__);
